Split district_inspect.rs (570 lines) into district_inspect/ subdirectory
Split crates/ui/src/district_inspect.rs into focused modules:
- mod.rs: thin facade with re-exports
- resources.rs: SelectedDistrict, DistrictPanelOpen, DistrictInspectCache
- helpers.rs: pure functions (grid_to_world_center, service_covers_cell, etc.)
- systems.rs: ECS systems and plugin definition
- ui_panel.rs: egui rendering
- tests.rs: unit tests

All public items re-exported from mod.rs — no external import changes needed.
